+++ tests/pipelines/chroma/test_pipeline_chroma.py
@@ -4,9 +4,9 @@ from transformers import AutoConfig, AutoTokenizer, T5EncoderModel
 from diffusers import AutoencoderKL, ChromaPipeline, ChromaTransformer2DModel, FlowMatchEulerDiscreteScheduler
 
 from ...testing_utils import assert_tensors_close, torch_device
-from ..flux.testing_utils import FluxIPAdapterTesterMixin
 from ..testing_utils import (
     BasePipelineTesterConfig,
+    FluxIPAdapterTesterMixin,
     MemoryTesterMixin,
     PipelineTesterMixin,
     check_qkv_fused_layers_exist,
