+++ tests/pipelines/controlnet_flux/test_controlnet_flux.py
@@ -40,8 +40,12 @@ from ...testing_utils import (
     require_big_accelerator,
     torch_device,
 )
-from ..flux.testing_utils import FluxIPAdapterTesterMixin
-from ..testing_utils import BasePipelineTesterConfig, MemoryTesterMixin, PipelineTesterMixin
+from ..testing_utils import (
+    BasePipelineTesterConfig,
+    FluxIPAdapterTesterMixin,
+    MemoryTesterMixin,
+    PipelineTesterMixin,
+)
 
 
 enable_full_determinism()
