+++ tests/pipelines/flux/test_pipeline_flux.py
@@ -31,6 +31,7 @@ from ..testing_utils import (
     BasePipelineTesterConfig,
     FasterCacheTesterMixin,
     FirstBlockCacheTesterMixin,
+    FluxIPAdapterTesterMixin,
     LoraMemoryTesterMixin,
     LoraTesterMixin,
     MagCacheTesterMixin,
@@ -39,7 +40,6 @@ from ..testing_utils import (
     PyramidAttentionBroadcastTesterMixin,
     TaylorSeerCacheTesterMixin,
 )
-from .testing_utils import FluxIPAdapterTesterMixin
 
 
 if is_peft_available():
