+++ tests/pipelines/flux/test_pipeline_flux_kontext.py
@@ -13,11 +13,11 @@ from ...testing_utils import torch_device
 from ..testing_utils import (
     BasePipelineTesterConfig,
     FasterCacheTesterMixin,
+    FluxIPAdapterTesterMixin,
     MemoryTesterMixin,
     PipelineTesterMixin,
     PyramidAttentionBroadcastTesterMixin,
 )
-from .testing_utils import FluxIPAdapterTesterMixin
 
 
 class FluxKontextPipelineTesterConfig(BasePipelineTesterConfig):
