+++ tests/pipelines/kandinsky/test_kandinsky.py
@@ -31,11 +31,11 @@ from ...testing_utils import (
     slow,
     torch_device,
 )
-from ..test_pipelines_common import assert_mean_pixel_difference
 from ..testing_utils import (
     BasePipelineTesterConfig,
     MemoryTesterMixin,
     PipelineTesterMixin,
+    assert_mean_pixel_difference,
 )
