+++ tests/pipelines/stable_diffusion_adapter/test_stable_diffusion_adapter.py
@@ -49,7 +49,8 @@ from ...testing_utils import (
     torch_device,
 )
 from ..pipeline_params import TEXT_GUIDED_IMAGE_VARIATION_BATCH_PARAMS, TEXT_GUIDED_IMAGE_VARIATION_PARAMS
-from ..test_pipelines_common import PipelineFromPipeTesterMixin, PipelineTesterMixin, assert_mean_pixel_difference
+from ..test_pipelines_common import PipelineFromPipeTesterMixin, PipelineTesterMixin
+from ..testing_utils import assert_mean_pixel_difference
 
 
 enable_full_determinism()
