+++ tests/pipelines/stable_unclip/test_stable_unclip_img2img.py
@@ -32,11 +32,11 @@ from ...testing_utils import (
     torch_device,
 )
 from ..pipeline_params import TEXT_GUIDED_IMAGE_VARIATION_BATCH_PARAMS, TEXT_GUIDED_IMAGE_VARIATION_PARAMS
-from ..test_pipelines_common import assert_mean_pixel_difference
 from ..testing_utils import (
     BasePipelineTesterConfig,
     MemoryTesterMixin,
     PipelineTesterMixin,
+    assert_mean_pixel_difference,
 )
