+++ tests/pipelines/test_pipelines_common.py
@@ -69,6 +69,7 @@ from ..testing_utils import (
     skip_mps,
     torch_device,
 )
+from .testing_utils import assert_mean_pixel_difference
 
 
 def to_np(tensor):
@@ -2934,13 +2935,3 @@ class MagCacheTesterMixin:
         assert np.allclose(original_image_slice, image_slice_disabled, atol=1e-4), (
             "Outputs after disabling cache should match original inference exactly."
         )
-
-
-# Some models (e.g. unCLIP) are extremely likely to significantly deviate depending on which hardware is used.
-# This helper function is used to check that the image doesn't deviate on average more than 10 pixels from a
-# reference image.
-def assert_mean_pixel_difference(image, expected_image, expected_max_diff=10):
-    image = np.asarray(DiffusionPipeline.numpy_to_pil(image)[0], dtype=np.float32)
-    expected_image = np.asarray(DiffusionPipeline.numpy_to_pil(expected_image)[0], dtype=np.float32)
-    avg_diff = np.abs(image - expected_image).mean()
-    assert avg_diff < expected_max_diff, f"Error image deviates {avg_diff} pixels on average"
