+++ tests/pipelines/testing_utils/__init__.py
@@ -8,7 +8,7 @@ from .cache import (
 )
 from .common import BasePipelineTesterConfig, PipelineTesterMixin
 from .from_pipe import FromPipeTesterMixin
-from .ip_adapter import IPAdapterTesterMixin
+from .ip_adapter import FluxIPAdapterTesterMixin, IPAdapterTesterMixin
 from .lora import LoraMemoryTesterMixin, LoraTesterMixin, UNetLoraTesterMixin
 from .memory import (
     GroupOffloadTesterMixin,
@@ -17,6 +17,7 @@ from .memory import (
     PipelineOffloadTesterMixin,
 )
 from .utils import (
+    assert_mean_pixel_difference,
     check_qkv_fused_layers_exist,
     check_qkv_fusion_matches_attn_procs_length,
     check_qkv_fusion_processors_exist,
@@ -29,6 +30,7 @@ __all__ = [
     "PipelineTesterMixin",
     "FromPipeTesterMixin",
     "IPAdapterTesterMixin",
+    "FluxIPAdapterTesterMixin",
     "LoraTesterMixin",
     "LoraMemoryTesterMixin",
     "UNetLoraTesterMixin",
@@ -42,6 +44,7 @@ __all__ = [
     "FirstBlockCacheTesterMixin",
     "TaylorSeerCacheTesterMixin",
     "MagCacheTesterMixin",
+    "assert_mean_pixel_difference",
     "check_qkv_fused_layers_exist",
     "check_qkv_fusion_matches_attn_procs_length",
     "check_qkv_fusion_processors_exist",
