+++ tests/pipelines/testing_utils/utils.py
@@ -13,6 +13,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import numpy as np
+
+from diffusers import DiffusionPipeline
 from diffusers.models.attention import AttentionModuleMixin
 
 
@@ -52,3 +55,13 @@ def check_qkv_fused_layers_exist(model, layer_names):
         is_fused = is_fused_attribute_set and is_fused_layer
         is_fused_submodules.append(is_fused)
     return all(is_fused_submodules)
+
+
+# Some models (e.g. unCLIP) are extremely likely to significantly deviate depending on which hardware is used.
+# This helper function is used to check that the image doesn't deviate on average more than 10 pixels from a
+# reference image.
+def assert_mean_pixel_difference(image, expected_image, expected_max_diff=10):
+    image = np.asarray(DiffusionPipeline.numpy_to_pil(image)[0], dtype=np.float32)
+    expected_image = np.asarray(DiffusionPipeline.numpy_to_pil(expected_image)[0], dtype=np.float32)
+    avg_diff = np.abs(image - expected_image).mean()
+    assert avg_diff < expected_max_diff, f"Error image deviates {avg_diff} pixels on average"
